Make recursive copies portable
On OSX, `cp -r` where the source directory name ends with a slash will
cause the contents of that dir (not the dir itself) to be copied to the
destination. This differs to the behaviour on Linux.

As a result the public and assets dirs do not appear with the correct
structure when building on OSX.

As a more portable alternative we can consistently use the pattern

    cp -R <src> <dest>/

with a preceding `mkdir` for <dest> if needed. This will create
`<dest>/<src>` on OSX and Linux.