fix: replace any backticks (`) in manifestStr
if manifestStr (read from plugin.json) has any '`' characters in it,
such as when using markdown formatting in a text value, the generated go
code fails to build due to unescaped '`' characters, leading the
compiler to believe some input is mismatched.

This change passes the manifestStr through a Replacement filter to
replace the ` characters with input which will be properly formatted
through the compiler.

Signed-off-by: Neil Hanlon <neil@shrug.pw>